services/weather_service.py
import os 
import requests
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("API_KEY")

def get_weather(city):
    url = "https://api.openweathermap.org/data/2.5/weather"
    params = {
        "q": city,
        "appid": API_KEY,
        "units": "metric"
    }

    response = requests.get(url, params=params)

    if response.status_code == 200:
        data = response.json()

        return { 
            "city": data["name"],
            "temperature": data["main"]["temp"],
            "condition": data["weather"][0]["main"]
        }
    
    return None